move decl to logger
